DollorBox/requirements.txt
asgiref==3.2.10
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.4
Django==3.1.1
django-environ==0.8.1
django-livereload==1.7
django-livereload-server==0.4
django-static-md5url==0.1
filelock==3.4.2
idna==3.3
jaconv==0.3
mysqlclient==2.1.0
numpy==1.22.0
oauthlib==3.1.1
pandas==1.3.5
Pillow==7.2.0
pipenv==2022.1.8
platformdirs==2.4.1
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
pykakasi==2.2.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.0
semidbm==0.5.1
six==1.16.0
social-auth-app-django==3.4.0
social-auth-core==3.3.3
soupsieve==2.3.1
sqlparse==0.4.2
tornado==6.1
urllib3==1.26.8
uWSGI==2.0.20
virtualenv==20.13.0
virtualenv-clone==0.5.7
wrapt==1.13.3
